refactor(editview): split up into more modules
The edit_view.rs files has become pretty huge, so it makes sense to split off
the GTK Widget stuff into view_item.rs and the font related stuff into fonts.rs
