+++ src/editview/src/lib.rs
@@ -2,8 +2,10 @@
 extern crate enclose;
 
 pub mod edit_view;
+pub mod fonts;
 pub mod main_state;
 pub mod theme;
+mod view_item;
 
 pub use crate::edit_view::EditView;
 pub use crate::main_state::{MainState, Settings};
